UI implemetation on Dashboard

+++ src/components/dashboard/ProductRow.jsx
@@ -1,5 +1,9 @@
 import React, { useState } from "react";
 import VariantsTable from "./VariantsTable"; // Asegúrate que esté el path correcto
+import eliminar from "../../images/eliminar.png";
+import guardar from "../../images/guardar.png";
+import editar from "../../images/editar.png";
+import Swal from "sweetalert2";
 
 export default function ProductRow({
   product,
@@ -63,12 +67,12 @@ export default function ProductRow({
       setError("Las variantes deben tener abreviación, nombre y precio.");
       return false;
     }
-    setError(""); // Sin errores
+    setError("");
     return true;
   };
 
   const handleSave = async () => {
-    if (!validateForm()) return; // Solo guarda si la validación pasa
+    if (!validateForm()) return;
     try {
       await updateProduct({ _id: product._id, ...editForm });
       setEditProductId(null);
@@ -79,48 +83,54 @@ export default function ProductRow({
   };
 
   const handleDelete = async () => {
-    try {
-      await deleteProduct(product._id);
-      await refreshProducts();
-    } catch (error) {
-      console.error(error);
+    const result = await Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Esto eliminará el producto permanentemente.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#018242",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    });
+
+    if (result.isConfirmed) {
+      try {
+        await deleteProduct(product._id);
+        await refreshProducts();
+        Swal.fire("Eliminado", "El producto ha sido eliminado.", "success");
+      } catch (error) {
+        console.error(error);
+        Swal.fire("Error", "No se pudo eliminar el producto.", "error");
+      }
     }
   };
 
   return (
-    <tr>
-      <td>
-        {isEditing ? (
-          <input
-            type="text"
-            name="name"
-            required
-            value={editForm.name}
-            onChange={handleChange}
-          />
-        ) : (
-          product.name
-        )}
+    <tr className="border-b-2 border-neutral-300">
+      <td className="pr-2 pt-1 align-top">
+        <div className="w-[120px] break-words">
+          {isEditing ? (
+            <input
+              type="text"
+              name="name"
+              required
+              className="w-full border rounded p-1 font-medium"
+              value={editForm.name}
+              onChange={handleChange}
+            />
+          ) : (
+            <p className="font-medium">{product.name}</p>
+            
+          )}
+        </div>
       </td>
-      <td>
-        {isEditing ? (
-          <input
-            type="number"
-            name="price"
-            required
-            min={1}
-            value={editForm.price}
-            onChange={handleChange}
-          />
-        ) : (
-          `₲${product.price}`
-        )}
-      </td>
-      <td>{new Date(product.createdAt).toLocaleDateString()}</td>
-      <td>
-        {isEditing
-          ? categories.map((cat) => (
-              <label key={cat._id} style={{ marginRight: "10px" }}>
+
+      <td className="align-top">
+        <div className="w-[170px] flex flex-wrap gap-2">
+          {isEditing ? (
+            categories.map((cat) => (
+              <label key={cat._id} className="flex items-center gap-1 text-xs">
                 <input
                   type="checkbox"
                   value={cat.name}
@@ -130,56 +140,105 @@ export default function ProductRow({
                 {cat.name}
               </label>
             ))
-          : product.category.join(", ")}
+          ) : (
+            <div className="flex flex-wrap ">
+              {product.category.map((cat) => (
+                <div
+                  key={cat}
+                  className=" mt-1 mr-1 p-1 rounded-lg bg-white"
+                  style={{ border: "2px solid #e0e0e0" }}
+                >
+                  <p>{cat}</p>
+                </div>
+              ))}
+            </div>
+          )}
+        </div>
       </td>
-      <td>
-        <VariantsTable
-          variants={isEditing ? editForm.variants : product.variants}
-          id={product._id}
-          editing={isEditing}
-          onEdit={(newVariants) =>
-            setEditForm({ ...editForm, variants: newVariants })
-          }
-        />
-      </td>
-      <td className="w-8">
-        {isEditing ? (
-          <input
-            type="text"
-            name="imageURL"
-            value={editForm.imageURL}
-            onChange={handleChange}
+
+      <td className="align-top">
+        <div className="min-w-[300px] overflow-x-auto">
+          <VariantsTable
+            variants={isEditing ? editForm.variants : product.variants}
+            id={product._id}
+            editing={isEditing}
+            onEdit={(newVariants) =>
+              setEditForm({ ...editForm, variants: newVariants })
+            }
           />
-        ) : (
-          product.imageURL
-        )}
+        </div>
       </td>
-      {error && <div style={{ color: "red", marginTop: "5px" }}>{error}</div>}
-      <td>
-        {isEditing ? (
-          <div>
-            <button onClick={handleSave}>Guardar</button>
-            <button onClick={() => setEditProductId(null)}>Cancelar</button>
-          </div>
-        ) : (
-          <>
-            <button
-              onClick={() => {
-                setEditProductId(product._id);
-                setEditForm({
-                  name: product.name,
-                  price: product.price,
-                  imageURL: product.imageURL,
-                  category: product.category,
-                  variants: product.variants,
-                });
-              }}
-            >
-              Editar
-            </button>
-            <button onClick={handleDelete}>Eliminar</button>
-          </>
-        )}
+
+      <td className="p-2 align-top">
+        <div className="min-w-[150px] break-all">
+          {isEditing ? (
+            <input
+              type="text"
+              name="imageURL"
+              className="w-full border rounded p-1"
+              value={editForm.imageURL}
+              onChange={handleChange}
+            />
+          ) : (
+            product.imageURL
+          )}
+        </div>
+      </td>
+
+      <td className="py-2 align-top">
+        <div className="min-w-[100px]">
+          {new Date(product.createdAt).toLocaleDateString()}
+        </div>
+      </td>
+
+      <td className="py-2 align-top">
+        <div className="w-[90px] flex flex-col gap-1">
+          {error && <p className="text-red-600 text-xs">{error}</p>}
+          {isEditing ? (
+            <div className="flex flex-col gap-1">
+              <button
+                onClick={handleSave}
+                className="bg-green-200 text-green-950 text-xs rounded px-2 py-1 flex items-center justify-between gap-1"
+              >
+                <p>Guardar</p>
+                <img className="w-4 h-4 object-contain" src={guardar} alt="" />
+              </button>
+              <button
+                onClick={()=>{setEditProductId(null)}}
+                className="bg-green-200 text-green-950 text-xs rounded px-2 py-1 flex items-center gap-1"
+              >
+                <span>Cancelar</span>
+                <img className="w-4 h-4 object-contain" src={eliminar} alt="" />
+              </button>
+            </div>
+          ) : (
+            <div className="flex flex-col gap-1">
+              <button
+                className="bg-green-200 text-green-950 text-xs rounded px-2 py-1 flex items-center gap-1"
+                onClick={() => {
+                  setEditProductId(product._id);
+                  setEditForm({
+                    name: product.name,
+                    price: product.price,
+                    imageURL: product.imageURL,
+                    category: product.category,
+                    variants: product.variants,
+                  });
+                }}
+              >
+                <span>Editar</span>
+                <img className="w-4 h-4 object-contain" src={editar} alt="" />
+              </button>
+              <button
+                onClick={handleDelete}
+                className="bg-green-200 text-green-950 text-xs rounded px-2 py-1 flex items-center gap-1"
+              >
+                <span>Eliminar</span>
+                <img className="w-4 h-4 object-contain" src={eliminar} alt="" />
+              </button>
+            </div>
+          )}
+        </div>
       </td>
     </tr>
   );
